ppfinjector#31: Add choose_emulator command to emulauncher
* Move the functions for launching emulator to their own file.
* Add choose_emulator command for choosing a different emulator to
  launch than the one stored in config.
