+++ app/emulauncher/src/launch_emulator.h
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace tdd::app::emulauncher {
+
+void LaunchEmulator();
+
+}
